AuditKeybinds: Light refactoring with some helpers. Refining parsing of clickabledata a little more.
