+++ CoverageReport/C-101EB.html
@@ -1,1 +1,1 @@
-<h1>SYSTEMS</h1><div>"Tap Glass"</div><div>"Tap Glass"</div><div>"HSI Digit Brightness"</div><div>"HSI Digit Brightness"</div><div>"ADI Pitch Adjust"</div><div>"ADI Pitch Adjust"</div><div>"GO-AROUND MODE"</div><div>"HEADING MODE"</div><div>"ALTITUDE HOLD MODE"</div><div>"PITCH MODE"</div><div>"VOR/LOC MODE"</div><div>"APPROACH ARM"</div><div>"GLIDESLOPE MODE"</div><div>"ILS/LOC REVERSE MODE"</div><div>"STANDBY MODE"</div><div>"GLIDESLOPE EXTENDED MODE"</div><div>"FD Bright/Dim"</div><div>"ADJUST RED LIGHT - TOP CENTER PANEL"</div><div>"ADJUST RED LIGHT - TOP CENTER PANEL"</div><div>"Manual Gyro Correction"</div><div>"Vertical Gyro Fast Erect"</div><div>"Gyro Mode"</div><div>"Manual Gyro Correction"</div><div>"Vertical Gyro Fast Erect"</div><div>"Gyro Mode"</div><div>"VOR/TCN Nav Selector"</div><div>"Airspeed Indicator Bug"</div><div>"VOR/TCN Nav Selector"</div><div>"Airspeed Indicator Bug"</div><div>"UHF ANTENNA SELECTION"</div><div>"VERTICAL GYRO DC"</div><div>"DIRECTIONAL GYRO DC"</div><div>"IFF PRIM DC"</div><div>"FUEL MEASURING SYSTEM"</div><div>"UHF"</div><div>"TACAN DC"</div><div>"RED COCKPIT LIGHTING"</div><div>"VHF DC ESS"</div><div>"INTERPHONE"</div><div>"BACKUP ADI"</div><div>"DIRECTIONAL GYRO AC"</div><div>"VERTICAL GYRO AC"</div><div>"TACAN AC 115"</div><div>"CONSOLE LIGHTS"</div><div>"IFF AC 115"</div><div>"FLIGHT DIRECTOR"</div><div>"TACAN AC 26"</div><div>"VOR AC 26"</div><div>"Colour Smoke"</div><div>"White Smoke"</div><h1>ARN127</h1><div>"TEST VOR/MKR"</div><div>"TEST VHF NAV DIGITS"</div><div>"DIMMER"</div><div>"108-117"</div><div>"Frequency selector (0-9)"</div><div>"Frequency selector (00/50)"</div><div>"NAV VOL"</div><div>"TEST VOR/MKR"</div><div>"TEST VHF NAV DIGITS"</div><div>"DIMMER"</div><div>"Frequency selector (108-117)"</div><div>"Frequency selector (0-9)"</div><div>"Frequency selector (00/50)"</div><div>"NAV VOL"</div><h1>ARC134</h1><div>"VHF Volume/Brightness"</div><div>"VHF Volume/Brightness"</div><div>"POWER SWITCH"</div><div>"DISPLAY TEST"</div><div>"Frequency selector (X0X.XXX)"</div><div>"Frequency selector (XX0.XXX)"</div><div>"Frequency selector (XXX.0XX)"</div><div>"Frequency selector (XXX.X00)"</div><div>"VHF Volume/Brightness"</div><div>"VHF Volume/Brightness"</div><div>"POWER SWITCH"</div><div>"DISPLAY TEST"</div><div>"Frequency selector (X0X.XXX)"</div><div>"Frequency selector (XX0.XXX)"</div><div>"Frequency selector (XXX.0XX)"</div><div>"Frequency selector (XXX.X00)"</div><h1>ARC164</h1><div>"Preset Channel Selector"</div><div>"100 MHz Selector"</div><div>"10 MHz Selector"</div><div>"1 MHz Selector"</div><div>"0.1 MHz Selector"</div><div>"0.025 MHz Selector"</div><div>"Frequency Mode Dial"</div><div>"Function Dial"</div><div>"T-Tone"</div><div>"Squelch Switch"</div><div>"Volume control"</div><div>"UHF Cover"</div><div>"UHF Preset Button"</div><div>"Preset Channel Selector"</div><div>"100 MHz Selector"</div><div>"10 MHz Selector"</div><div>"1 MHz Selector"</div><div>"0.1 MHz Selector"</div><div>"0.025 MHz Selector"</div><div>"Frequency Mode Dial"</div><div>"Function Dial"</div><div>"T-Tone"</div><div>"Squelch Switch"</div><div>"Volume control"</div><div>"UHF Cover"</div><div>"UHF Preset Button"</div><div>"UHF Digital Display Dimmer"</div><div>"UHF Digital Display Dimmer"</div><h1>Button_13</h1><div></div><h1>Button_38</h1><div></div><h1>INTERCOM</h1><div>"UHF Switch"</div><div>"TCN Switch"</div><div>"Selector switch"</div><div>"UHF Switch"</div><div>"TCN Switch"</div><div>"Selector switch"</div><h1>ARN118</h1><div>"00x-12x. Rotate mouse wheel to select"</div><div>"Channel Selector (Ones) / X/Y Mode. Right mouse click to select X/Y. Rotate mouse wheel to make channel selection"</div><div>"Channel Selector (Ones) / X/Y Mode. Right mouse click to select X/Y. Rotate mouse wheel to make channel selection"</div><div>"TACAN TEST"</div><div>"TACAN Volume"</div><div>"TACAN Mode"</div>
+<h1>SYSTEMS</h1><div>"Tap Glass"</div><div>"Tap Glass"</div><div>"HSI Digit Brightness"</div><div>"HSI Digit Brightness"</div><div>"ADI Pitch Adjust"</div><div>"ADI Pitch Adjust"</div><div>"GO-AROUND MODE"</div><div>"HEADING MODE"</div><div>"ALTITUDE HOLD MODE"</div><div>"PITCH MODE"</div><div>"VOR/LOC MODE"</div><div>"APPROACH ARM"</div><div>"GLIDESLOPE MODE"</div><div>"ILS/LOC REVERSE MODE"</div><div>"STANDBY MODE"</div><div>"GLIDESLOPE EXTENDED MODE"</div><div>"FD Bright/Dim"</div><div>"ADJUST RED LIGHT - TOP CENTER PANEL"</div><div>"ADJUST RED LIGHT - TOP CENTER PANEL"</div><div>"Manual Gyro Correction"</div><div>"Vertical Gyro Fast Erect"</div><div>"Gyro Mode"</div><div>"Manual Gyro Correction"</div><div>"Vertical Gyro Fast Erect"</div><div>"Gyro Mode"</div><div>"VOR/TCN Nav Selector"</div><div>"Airspeed Indicator Bug"</div><div>"VOR/TCN Nav Selector"</div><div>"Airspeed Indicator Bug"</div><div>"UHF ANTENNA SELECTION"</div><div>"VERTICAL GYRO DC"</div><div>"DIRECTIONAL GYRO DC"</div><div>"IFF PRIM DC"</div><div>"FUEL MEASURING SYSTEM"</div><div>"UHF"</div><div>"TACAN DC"</div><div>"RED COCKPIT LIGHTING"</div><div>"VHF DC ESS"</div><div>"INTERPHONE"</div><div>"BACKUP ADI"</div><div>"DIRECTIONAL GYRO AC"</div><div>"VERTICAL GYRO AC"</div><div>"TACAN AC 115"</div><div>"CONSOLE LIGHTS"</div><div>"IFF AC 115"</div><div>"FLIGHT DIRECTOR"</div><div>"TACAN AC 26"</div><div>"VOR AC 26"</div><div>"Colour Smoke"</div><div>"White Smoke"</div><h1>ARN127</h1><div>"TEST VOR/MKR"</div><div>"TEST VHF NAV DIGITS"</div><div>"DIMMER"</div><div>"108-117"</div><div>"Frequency selector (0-9)"</div><div>"Frequency selector (00/50)"</div><div>"NAV VOL"</div><div>"TEST VOR/MKR"</div><div>"TEST VHF NAV DIGITS"</div><div>"DIMMER"</div><div>"Frequency selector (108-117)"</div><div>"Frequency selector (0-9)"</div><div>"Frequency selector (00/50)"</div><div>"NAV VOL"</div><h1>ARC134</h1><div>"POWER SWITCH"</div><div>"DISPLAY TEST"</div><div>"Frequency selector (X0X.XXX)"</div><div>"Frequency selector (XX0.XXX)"</div><div>"Frequency selector (XXX.0XX)"</div><div>"Frequency selector (XXX.X00)"</div><div>"POWER SWITCH"</div><div>"DISPLAY TEST"</div><div>"Frequency selector (X0X.XXX)"</div><div>"Frequency selector (XX0.XXX)"</div><div>"Frequency selector (XXX.0XX)"</div><div>"Frequency selector (XXX.X00)"</div><h1>ARC164</h1><div>"Preset Channel Selector"</div><div>"100 MHz Selector"</div><div>"10 MHz Selector"</div><div>"1 MHz Selector"</div><div>"0.1 MHz Selector"</div><div>"0.025 MHz Selector"</div><div>"Frequency Mode Dial"</div><div>"Function Dial"</div><div>"T-Tone"</div><div>"Squelch Switch"</div><div>"Volume control"</div><div>"UHF Cover"</div><div>"UHF Preset Button"</div><div>"Preset Channel Selector"</div><div>"100 MHz Selector"</div><div>"10 MHz Selector"</div><div>"1 MHz Selector"</div><div>"0.1 MHz Selector"</div><div>"0.025 MHz Selector"</div><div>"Frequency Mode Dial"</div><div>"Function Dial"</div><div>"T-Tone"</div><div>"Squelch Switch"</div><div>"Volume control"</div><div>"UHF Cover"</div><div>"UHF Preset Button"</div><div>"UHF Digital Display Dimmer"</div><div>"UHF Digital Display Dimmer"</div><h1>Button_13</h1><div></div><h1>Button_38</h1><div></div><h1>INTERCOM</h1><div>"UHF Switch"</div><div>"TCN Switch"</div><div>"Selector switch"</div><div>"UHF Switch"</div><div>"TCN Switch"</div><div>"Selector switch"</div><h1>ARN118</h1><div>"TACAN TEST"</div><div>"TACAN Volume"</div><div>"TACAN Mode"</div>
