+++ CoverageReport/Christen Eagle II.html
@@ -1,1 +1,1 @@
-<h1>GENERAL</h1><div>"Accelerometer Reset Button"</div><div>"Smoke"</div><div>"Air Conditioning"</div><div>"Trim Lever"</div><h1>PITOT</h1><div>"Barometric Pressure"</div><h1>LIGHTS</h1><div>"Nav Lights"</div><div>"Landing Light"</div><div>"Cockpit Light"</div><div>"Cockpit Light Red"</div><div>"Cockpit Light Yellow"</div><h1>RADIO</h1><div>"Radio ON-OFF"</div><div>"Radio Frequency MHz"</div><div>"Radio Frequency 50KHz / 25KHz Fine Tune (Pull)"</div><div>"Radio Frequency Transfer"</div><div>"Radio Channel"</div><div>"Radio ON-OFF / Volume / Test (Pull)"</div><h1>ENGINE</h1><div>"Magnetos"</div><div>"Fuel Valve"</div><div>"Fuel Pump"</div><div>"Mixture"</div><div>"Propeller"</div><h1>ELECTRIC</h1><div>"Battery"</div><div>"Alternator"</div><div>"Fuse"</div><div></div><h1>FUELGAUGE</h1><div>"Fuel Indicator Mode (F / U)"</div><div>"Fuel Indicator Mode Rear cockpit(F / U)"</div><h1>CANOPY</h1><div>"Canopy Lever Open (Rear)"</div><div>"Canopy Lever Close (Rear)"</div><div>"Canopy Lever Open (Front)"</div><div>"Canopy Lock (Rear)"</div><div>"Canopy Unlock (Rear)"</div><div>"Canopy Lock (Front)"</div><div>"Canopy Unlock (Front)"</div><div>"Canopy Jettison (Rear)"</div>
+<h1>GENERAL</h1><div>"Accelerometer Reset Button"</div><div>"Smoke"</div><div>"Air Conditioning"</div><div>"Trim Lever"</div><h1>PITOT</h1><div>"Barometric Pressure"</div><h1>LIGHTS</h1><div>"Nav Lights"</div><div>"Landing Light"</div><div>"Cockpit Light"</div><div>"Cockpit Light Red"</div><div>"Cockpit Light Yellow"</div><h1>RADIO</h1><div>"Radio ON-OFF"</div><div>"Radio Frequency MHz"</div><div>"Radio Frequency 50KHz / 25KHz Fine Tune (Pull)"</div><div>"Radio Frequency Transfer"</div><div>"Radio Channel"</div><div>"Radio ON-OFF / Volume / Test (Pull)"</div><h1>ENGINE</h1><div>"Magnetos"</div><div>"Fuel Valve"</div><div>"Fuel Pump"</div><div>"Mixture"</div><div>"Propeller"</div><h1>ELECTRIC</h1><div>"Battery"</div><div>"Alternator"</div><div>"Fuse"</div><div>"Volt/Ampere Meter mode (Voltage / Amperage)"</div><h1>FUELGAUGE</h1><div>"Fuel Indicator Mode (F / U)"</div><div>"Fuel Indicator Mode Rear cockpit(F / U)"</div><h1>CANOPY</h1><div>"Canopy Lever Open (Rear)"</div><div>"Canopy Lever Close (Rear)"</div><div>"Canopy Lever Open (Front)"</div><div>"Canopy Lock (Rear)"</div><div>"Canopy Unlock (Rear)"</div><div>"Canopy Lock (Front)"</div><div>"Canopy Unlock (Front)"</div><div>"Canopy Jettison (Rear)"</div>
