+++ CoverageReport/FW-190D9.html
@@ -1,1 +1,1 @@
-<h1>CONTROL_SYSTEM</h1><div>"Radiator Flaps Control"</div><div>"Flaps Up"</div><div>"Flaps Take Off"</div><div>"Flaps Landing"</div><div>"Landing Gears Retraction"</div><div>"Landing Gears Retraction Cover"</div><div>"Landing Gears Extending"</div><div>"Stabilizer Trimmer Up/Down"</div><div>"Landing Gear Emergency Release Handle"</div><h1>FUEL_SYSTEM</h1><div>"Cold Start/Wind Screen Washer"</div><div>"Fuel Tank Selector Valve (CLOSE/FORWARD/AFT/BOTH)"</div><div>"Fuel Gauge Selector (FORWARD/NONE/AFT)"</div><div>"MW-B4 Selector (Not Functional)"</div><div>"Primer Pump"</div><h1>ENGINE_SYSTEM</h1><div>"Magneto Switch"</div><div>"Starter Switch Cover"</div><div>"Starter Switch. Left Button - Starter Power. Right Button - Magnetic Clutch"</div><div>"Starter Switch. Left Button - Starter Power. Right Button - Magnetic Clutch"</div><div>"MBG Emergency Mode Handle"</div><div>"MW-50 Switch"</div><div>"Throttle Lock"</div><h1>ELEC_INTERFACE</h1><div>"Electric Kill-switch"</div><div>"Emergency Equipment Destruction"</div><div>"Circuit Breakers Cover"</div><h1>CPT_MECH</h1><div>"Canopy Emergency Release Handle"</div><div>"Canopy Hand Crank"</div><h1>ALTIMETER</h1><div>"Altimeter Pressure Set"</div><h1>ARTIFICIAL_HORIZON</h1><div>"Horizon Cage"</div><h1>REMOTE_COMPASS</h1><div>"Course Set"</div><h1>CLOCK</h1><div>"Cockpit.Bf109K4.clock_scale"</div><div>"Cockpit.Bf109K4.clock_pinion"</div><div>"Cockpit.Bf109K4.clock_stop"</div><div>"Cockpit.Bf109K4.clock_stopwatch"</div><h1>WEAPON_SYSTEM</h1><div>"Master Arm"</div><div>"Set Gun 1 Ammunition Count"</div><div>"Set Gun 2 Ammunition Count"</div><div>"Set Gun 3 Ammunition Count"</div><div>"Set Gun 4 Ammunition Count"</div><div>"Main Rocket Switch"</div><div>"Rocket Emergency Release"</div><div>"Bomb Fusing Selector"</div><div>"Jettison Fuselage Stores"</div><h1>EZ42_GUNSIGHT</h1><div>"EZ42 Gunsight Gyro Power Switch, ON/OFF"</div><div>"EZ42 Gunsight Target Wingspan Knob"</div><div>"EZ42 Gunsight Night Lens Filter, ON/OFF"</div><div>"EZ42 Gunsight Target Distance Control"</div><div>"EZ42 Gunsight Brightness Control"</div><div>"EZ42 Gunsight Altitude Aiming Correction Adjusting Knob I"</div><h1>LIGHT_SYSTEM</h1><div>"Instrument Lights Brightness"</div><h1>VHF_RADIO</h1><div>"Radio Channel Selector"</div><div>"Radio Volume"</div><div>"FT FT / Y ZF Mode Switch"</div><div>"Radio Tuning (+/-30 kHz)"</div><h1>OXYGEN_SYSTEM</h1><div>"Oxygen Flow Valve"</div><div>"Oxygen Emergency Knob"</div><h1>IFF</h1><div>"IFF Channel Selector"</div><div>"IFF Check"</div>
+<h1>CONTROL_SYSTEM</h1><div>"Radiator Flaps Control"</div><div>"Flaps Up"</div><div>"Flaps Take Off"</div><div>"Flaps Landing"</div><div>"Landing Gears Retraction"</div><div>"Landing Gears Retraction Cover"</div><div>"Landing Gears Extending"</div><div>"Stabilizer Trimmer Up/Down"</div><div>"Landing Gear Emergency Release Handle"</div><h1>FUEL_SYSTEM</h1><div>"Cold Start/Wind Screen Washer"</div><div>"Fuel Tank Selector Valve (CLOSE/FORWARD/AFT/BOTH)"</div><div>"Fuel Gauge Selector (FORWARD/NONE/AFT)"</div><div>"MW-B4 Selector (Not Functional)"</div><h1>ENGINE_SYSTEM</h1><div>"Magneto Switch"</div><div>"Starter Switch Cover"</div><div>"MBG Emergency Mode Handle"</div><div>"MW-50 Switch"</div><div>"Throttle Lock"</div><h1>ELEC_INTERFACE</h1><div>"Electric Kill-switch"</div><div>"Emergency Equipment Destruction"</div><div>"Circuit Breakers Cover"</div><h1>CPT_MECH</h1><div>"Canopy Emergency Release Handle"</div><div>"Canopy Hand Crank"</div><h1>ALTIMETER</h1><div>"Altimeter Pressure Set"</div><h1>ARTIFICIAL_HORIZON</h1><div>"Horizon Cage"</div><h1>REMOTE_COMPASS</h1><div>"Course Set"</div><h1>CLOCK</h1><div>"Cockpit.Bf109K4.clock_scale"</div><div>"Cockpit.Bf109K4.clock_pinion"</div><div>"Cockpit.Bf109K4.clock_stop"</div><div>"Cockpit.Bf109K4.clock_stopwatch"</div><h1>WEAPON_SYSTEM</h1><div>"Master Arm"</div><div>"Set Gun 1 Ammunition Count"</div><div>"Set Gun 2 Ammunition Count"</div><div>"Set Gun 3 Ammunition Count"</div><div>"Set Gun 4 Ammunition Count"</div><div>"Main Rocket Switch"</div><div>"Rocket Emergency Release"</div><div>"Bomb Fusing Selector"</div><div>"Jettison Fuselage Stores"</div><h1>EZ42_GUNSIGHT</h1><div>"EZ42 Gunsight Gyro Power Switch, ON/OFF"</div><div>"EZ42 Gunsight Target Wingspan Knob"</div><div>"EZ42 Gunsight Night Lens Filter, ON/OFF"</div><div>"EZ42 Gunsight Target Distance Control"</div><div>"EZ42 Gunsight Brightness Control"</div><div>"EZ42 Gunsight Altitude Aiming Correction Adjusting Knob I"</div><h1>LIGHT_SYSTEM</h1><div>"Instrument Lights Brightness"</div><h1>VHF_RADIO</h1><div>"Radio Channel Selector"</div><div>"Radio Volume"</div><div>"FT FT / Y ZF Mode Switch"</div><div>"Radio Tuning (+/-30 kHz)"</div><h1>OXYGEN_SYSTEM</h1><div>"Oxygen Flow Valve"</div><div>"Oxygen Emergency Knob"</div><h1>IFF</h1><div>"IFF Channel Selector"</div><div>"IFF Check"</div>
