+++ CoverageReport/I-16.html
@@ -1,1 +1,1 @@
-<h1>ELECTRIC_SYSTEM</h1><div></div><div></div><div></div><div></div><div></div><div>"Radio"</div><div></div><div></div><div></div><div></div><div>"Sight dimmer rheostat"</div><div>"Cockpit light rheostat"</div><h1>GEAR_SYSTEM</h1><div>"Gear lock"</div><div>"Gear direction"</div><div>"Brake spring handle"</div><div>"Gear handle"</div><h1>FLAPS_SYSTEM</h1><div>"Flaps handle"</div><h1>MOTOR_SYSTEM</h1><div></div><div></div><div></div>
+<h1>ELECTRIC_SYSTEM</h1><div>"Clock heat"</div><div>"Pitot heat"</div><div>"Nav light"</div><div>"Accumulator"</div><div>"Shunt"</div><div>"Radio"</div><div>"Fire left torch"</div><div>"Torches jettison"</div><div>"Fire right torch"</div><div>"Nav light test"</div><div>"Sight dimmer rheostat"</div><div>"Cockpit light rheostat"</div><h1>GEAR_SYSTEM</h1><div>"Gear lock"</div><div>"Gear direction"</div><div>"Brake spring handle"</div><div>"Gear handle"</div><h1>FLAPS_SYSTEM</h1><div>"Flaps handle"</div><h1>MOTOR_SYSTEM</h1><div>"Magneto"</div><div>"Motor shutdown"</div><div>"Ratchet relay"</div><div>"Primary fuel switch"</div><div>"Start pump"</div><div>"WEP"</div><div>"RPM"</div><div>"Mixture"</div><div>"Throttle"</div><div>"Supercharge speed 1 - 0 - 2"</div><div>"Motor cowl flaps"</div><div>"Oil radiator"</div><div>"Fuel valve"</div><div>"Manual ratchet relay"</div><div>"External fuel tank valve"</div><h1>WEAPON_SYSTEM</h1><div>"Gun wing left recharge"</div><div>"Gun wing right recharge"</div><div>"Gun nose left recharge"</div><div>"Gun nose right recharge"</div><div>"Rockets/bombs counter"</div><div>"Rockets/bombs master arm"</div><div>"Jettison fuel tank"</div><h1>COMMON_SYSTEM</h1><div>"Course cage knob"</div><div>"Seat level"</div><div>"Set pressure"</div><div>"Left door"</div><div>"Right door"</div><div>"Mech sight"</div><div>"Sight filter"</div><h1>OXYGEN_SYSTEM</h1><div>"Oxygen adjusting valve"</div><div>"Oxygen emergency valve"</div><div>"Oxygen shutoff valve"</div>
