--- CoverageReport/Ka-50.html
@@ -1,1 +1,0 @@
-<h1>CPT_MECH</h1><div>"Gear lever"</div><div>"Gear lever"</div><div>"Main/Common hydraulics for emergency gear"</div><div>"Main/Common hydraulics for emergency gear"</div><div>"Main/Common hydraulics for emergency gear"</div><div>"Main/Common hydraulics for emergency gear"</div><h1>ACCELEROMETER</h1><div>"Accelerometer reset"</div><h1>SYST_CONTROLLER</h1><div>"Master caution push-light"</div><div>"Lamps test"</div><div>"Rotor RPM warning push-light"</div><div>"MWS operation mode (not functional)"</div><div>"MWS operation mode (not functional)"</div><h1>ELEC_INTERFACE</h1><div>"DC ground power"</div><div>"DC ground power"</div><div>"DC ground power"</div><div>"DC ground power"</div><div>"Battery 2"</div><div>"Battery 2"</div><div>"Battery 2"</div><div>"Battery 2"</div><div>"Battery 1"</div><div>"Battery 1"</div><div>"Battery 1"</div><div>"Battery 1"</div><div>"AC ground power"</div><div>"AC ground power"</div><div>"AC left generator"</div><div>"AC left generator"</div><div>"AC right generator"</div><div>"AC right generator"</div><div>"DC/AC inverter"</div><div>"DC/AC inverter"</div><h1>FUELSYS_INTERFACE</h1><div>"Forward fuel tank pumps"</div><div>"Forward fuel tank pumps"</div><div>"Rear fuel tank pumps"</div><div>"Rear fuel tank pumps"</div><div>"Inner external fuel tanks pumps"</div><div>"Inner external fuel tanks pumps"</div><div>"Outer external fuel tanks pumps"</div><div>"Outer external fuel tanks pumps"</div><div>"Fuelmeter power"</div><div>"Fuelmeter power"</div><div>"Left engine fuel shutoff valve"</div><div>"Left engine fuel shutoff valve"</div><div>"Left engine fuel shutoff valve"</div><div>"Left engine fuel shutoff valve"</div><div>"Right engine fuel shutoff valve"</div><div>"Right engine fuel shutoff valve"</div><div>"Right engine fuel shutoff valve"</div><div>"Right engine fuel shutoff valve"</div><div>"APU fuel shutoff valve"</div><div>"APU fuel shutoff valve"</div><div>"APU fuel shutoff valve"</div><div>"APU fuel shutoff valve"</div><div>"Fuel crossfeed valve"</div><div>"Fuel crossfeed valve"</div><div>"Fuel crossfeed valve"</div><div>"Fuel crossfeed valve"</div><div>"Fuel quantity indicator self test button"</div><h1>SPU_9</h1><div>"Intercom"</div><div>"Intercom"</div><h1>R_828</h1><div>"VHF-1 (R828) power switch"</div><div>"VHF-1 (R828) power switch"</div><h1>R_800</h1><div>"VHF-2 (R-800) power switch"</div><div>"VHF-2 (R-800) power switch"</div><h1>DATALINK</h1><div>"Radio equipment datalink TLK power switch"</div><div>"Radio equipment datalink TLK power switch"</div><div>"Radio equipment datalink UHF TLK power switch"</div><div>"Radio equipment datalink UHF TLK power switch"</div>
