+++ CoverageReport/SpitfireLFMkIX.html
@@ -1,1 +1,1 @@
-<h1>OXYGEN</h1><div>"Cockpit.SpitfireLFMkIX.oxygen_valve"</div><h1>WEAPONS</h1><div>"Cockpit.SpitfireLFMkIX.safety"</div><div>"Cockpit.SpitfireLFMkIX.trigger_0"</div><div>"Cockpit.SpitfireLFMkIX.trigger_1"</div><div>"Cockpit.SpitfireLFMkIX.trigger_2"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_range"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_span"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_tint"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_master"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_illumination"</div><div>"Cockpit.SpitfireLFMkIX.trigger_bomb"</div><div>"Cockpit.SpitfireLFMkIX.droptank_release"</div><h1>CONTROLS</h1><div>"Cockpit.SpitfireLFMkIX.wheel_brakes"</div><div>"Cockpit.SpitfireLFMkIX.altimeter"</div><div>"Cockpit.SpitfireLFMkIX.di"</div><div>"Cockpit.SpitfireLFMkIX.fuel_gauge"</div><div>"Cockpit.SpitfireLFMkIX.nav_lights"</div><div>"Cockpit.SpitfireLFMkIX.flaps"</div><div>"Cockpit.SpitfireLFMkIX.uc_blind"</div><div>"Cockpit.SpitfireLFMkIX.compass"</div><div>"Cockpit.SpitfireLFMkIX.trim_elevator"</div><div>"Cockpit.SpitfireLFMkIX.trim_rudder"</div><div>"Cockpit.SpitfireLFMkIX.radiator_mode"</div><div>"Cockpit.SpitfireLFMkIX.pitot"</div><div>"Cockpit.SpitfireLFMkIX.deicer"</div><div>"Cockpit.SpitfireLFMkIX.uc_emergency"</div><div>"Cockpit.SpitfireLFMkIX.uc"</div><div>"Cockpit.SpitfireLFMkIX.canopy_operate"</div><div>"Cockpit.SpitfireLFMkIX.canopy_jettison"</div><div>"Cockpit.SpitfireLFMkIX.sidedoor_operate"</div><h1>CLOCK</h1><div>"Cockpit.SpitfireLFMkIX.clock"</div><div>"Cockpit.SpitfireLFMkIX.clock"</div><h1>ENGINE_CONTROLS</h1><div>"Cockpit.SpitfireLFMkIX.magneto0"</div><div>"Cockpit.SpitfireLFMkIX.magneto1"</div><div>"Cockpit.SpitfireLFMkIX.supercharger_mode"</div><div>"Cockpit.SpitfireLFMkIX.starter_cover"</div><div>"Cockpit.SpitfireLFMkIX.starter"</div><div>"Cockpit.SpitfireLFMkIX.booster_cover"</div><div>"Cockpit.SpitfireLFMkIX.booster"</div><div>"Cockpit.SpitfireLFMkIX.primer"</div><div>"Cockpit.SpitfireLFMkIX.primer"</div><div>"Cockpit.SpitfireLFMkIX.tank_pressurizer"</div><div>"Cockpit.SpitfireLFMkIX.throttle"</div><div>"Cockpit.SpitfireLFMkIX.pitch"</div><div>"Cockpit.SpitfireLFMkIX.mix"</div><div>"Cockpit.SpitfireLFMkIX.uc_down_indication"</div><div>"Cockpit.SpitfireLFMkIX.fuel_pump"</div><div>"Cockpit.SpitfireLFMkIX.carburettor_flap"</div><div>"Cockpit.SpitfireLFMkIX.dilution_cover"</div><div>"Cockpit.SpitfireLFMkIX.dilution"</div><div>"Cockpit.SpitfireLFMkIX.supercharger_cover"</div><div>"Cockpit.SpitfireLFMkIX.supercharger_test"</div><div>"Cockpit.SpitfireLFMkIX.radiator_cover"</div><div>"Cockpit.SpitfireLFMkIX.radiator_test"</div><div>"Cockpit.SpitfireLFMkIX.wobble"</div><div>"Cockpit.SpitfireLFMkIX.fuel_cock_primary"</div><div>"Cockpit.SpitfireLFMkIX.fuel_cock_droptank"</div><h1>TERTIARY_CONTROLS</h1><div>"Cockpit.SpitfireLFMkIX.illumination_lh"</div><div>"Cockpit.SpitfireLFMkIX.illumination_rh"</div><div>"Cockpit.SpitfireLFMkIX.id_lamp_up_mode"</div><div>"Cockpit.SpitfireLFMkIX.id_lamp_down_mode"</div><div>"Cockpit.SpitfireLFMkIX.morse_key"</div><div>"Cockpit.SpitfireLFMkIX.iff_b"</div><div>"Cockpit.SpitfireLFMkIX.iff_d"</div><div>"Cockpit.SpitfireLFMkIX.iff_cover"</div><div>"Cockpit.SpitfireLFMkIX.iff_0"</div><div>"Cockpit.SpitfireLFMkIX.iff_1"</div><h1>VHF_RADIO</h1><div>"Cockpit.SpitfireLFMkIX.radio_0"</div><div>"Cockpit.SpitfireLFMkIX.radio_a"</div><div>"Cockpit.SpitfireLFMkIX.radio_b"</div><div>"Cockpit.SpitfireLFMkIX.radio_c"</div><div>"Cockpit.SpitfireLFMkIX.radio_d"</div><div>"Cockpit.SpitfireLFMkIX.radio_dimmer"</div><div>"Cockpit.SpitfireLFMkIX.radio_lock"</div><div>"Cockpit.SpitfireLFMkIX.radio_mode"</div>
+<h1>OXYGEN</h1><div>"Cockpit.SpitfireLFMkIX.oxygen_valve"</div><h1>WEAPONS</h1><div>"Cockpit.SpitfireLFMkIX.safety"</div><div>"Cockpit.SpitfireLFMkIX.trigger_0"</div><div>"Cockpit.SpitfireLFMkIX.trigger_1"</div><div>"Cockpit.SpitfireLFMkIX.trigger_2"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_range"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_span"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_tint"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_master"</div><div>"Cockpit.SpitfireLFMkIX.gun_sight_illumination"</div><div>"Cockpit.SpitfireLFMkIX.trigger_bomb"</div><div>"Cockpit.SpitfireLFMkIX.droptank_release"</div><h1>CONTROLS</h1><div>"Cockpit.SpitfireLFMkIX.wheel_brakes"</div><div>"Cockpit.SpitfireLFMkIX.altimeter"</div><div>"Cockpit.SpitfireLFMkIX.di"</div><div>"Cockpit.SpitfireLFMkIX.fuel_gauge"</div><div>"Cockpit.SpitfireLFMkIX.nav_lights"</div><div>"Cockpit.SpitfireLFMkIX.flaps"</div><div>"Cockpit.SpitfireLFMkIX.uc_blind"</div><div>"Cockpit.SpitfireLFMkIX.compass"</div><div>"Cockpit.SpitfireLFMkIX.trim_elevator"</div><div>"Cockpit.SpitfireLFMkIX.trim_rudder"</div><div>"Cockpit.SpitfireLFMkIX.radiator_mode"</div><div>"Cockpit.SpitfireLFMkIX.pitot"</div><div>"Cockpit.SpitfireLFMkIX.deicer"</div><div>"Cockpit.SpitfireLFMkIX.uc_emergency"</div><div>"Cockpit.SpitfireLFMkIX.uc"</div><div>"Cockpit.SpitfireLFMkIX.canopy_operate"</div><div>"Cockpit.SpitfireLFMkIX.canopy_jettison"</div><div>"Cockpit.SpitfireLFMkIX.sidedoor_operate"</div><h1>ENGINE_CONTROLS</h1><div>"Cockpit.SpitfireLFMkIX.magneto0"</div><div>"Cockpit.SpitfireLFMkIX.magneto1"</div><div>"Cockpit.SpitfireLFMkIX.supercharger_mode"</div><div>"Cockpit.SpitfireLFMkIX.starter_cover"</div><div>"Cockpit.SpitfireLFMkIX.starter"</div><div>"Cockpit.SpitfireLFMkIX.booster_cover"</div><div>"Cockpit.SpitfireLFMkIX.booster"</div><div>"Cockpit.SpitfireLFMkIX.tank_pressurizer"</div><div>"Cockpit.SpitfireLFMkIX.throttle"</div><div>"Cockpit.SpitfireLFMkIX.pitch"</div><div>"Cockpit.SpitfireLFMkIX.mix"</div><div>"Cockpit.SpitfireLFMkIX.uc_down_indication"</div><div>"Cockpit.SpitfireLFMkIX.fuel_pump"</div><div>"Cockpit.SpitfireLFMkIX.carburettor_flap"</div><div>"Cockpit.SpitfireLFMkIX.dilution_cover"</div><div>"Cockpit.SpitfireLFMkIX.dilution"</div><div>"Cockpit.SpitfireLFMkIX.supercharger_cover"</div><div>"Cockpit.SpitfireLFMkIX.supercharger_test"</div><div>"Cockpit.SpitfireLFMkIX.radiator_cover"</div><div>"Cockpit.SpitfireLFMkIX.radiator_test"</div><div>"Cockpit.SpitfireLFMkIX.wobble"</div><div>"Cockpit.SpitfireLFMkIX.fuel_cock_primary"</div><div>"Cockpit.SpitfireLFMkIX.fuel_cock_droptank"</div><h1>TERTIARY_CONTROLS</h1><div>"Cockpit.SpitfireLFMkIX.illumination_lh"</div><div>"Cockpit.SpitfireLFMkIX.illumination_rh"</div><div>"Cockpit.SpitfireLFMkIX.id_lamp_up_mode"</div><div>"Cockpit.SpitfireLFMkIX.id_lamp_down_mode"</div><div>"Cockpit.SpitfireLFMkIX.morse_key"</div><div>"Cockpit.SpitfireLFMkIX.iff_b"</div><div>"Cockpit.SpitfireLFMkIX.iff_d"</div><div>"Cockpit.SpitfireLFMkIX.iff_cover"</div><div>"Cockpit.SpitfireLFMkIX.iff_0"</div><div>"Cockpit.SpitfireLFMkIX.iff_1"</div><h1>VHF_RADIO</h1><div>"Cockpit.SpitfireLFMkIX.radio_0"</div><div>"Cockpit.SpitfireLFMkIX.radio_a"</div><div>"Cockpit.SpitfireLFMkIX.radio_b"</div><div>"Cockpit.SpitfireLFMkIX.radio_c"</div><div>"Cockpit.SpitfireLFMkIX.radio_d"</div><div>"Cockpit.SpitfireLFMkIX.radio_dimmer"</div><div>"Cockpit.SpitfireLFMkIX.radio_lock"</div><div>"Cockpit.SpitfireLFMkIX.radio_mode"</div>
